3.0.2 - To small to name it.
Changes
+ Changed the Update notifications tab from /profile.
+ Changed the "No Permission" message for message components.
+ Changed the message sent when you add QuaBot.

Bug Fixes
- Console spam on /clear.
- Crashes when some permissions were missing.

Dev Changes
- Bumped version to 3.0.2
- Changed token to Quabot testing
- Added logging when added & removed from guild.

+++ commands/management/clear.js
@@ -23,6 +23,13 @@ module.exports = {
             let amount = interaction.options.getInteger('amount');
             let public = !interaction.options.getBoolean('private');
 
+            
+            if (interaction.channel.type === "GUILD_CATEGORY") return;
+            if (interaction.channel.type === "GUILD_DIRECTORY") return;
+            if (interaction.channel.type === "GUILD_FORUM") return;
+            if (interaction.channel.type === "GUILD_VOICE") return;
+            if (interaction.channel.type === "GUILD_STAGE_VOICE") return;
+
             if (amount > 0) {
                 if (amount > 100) return interaction.reply({ content: `You can't delete more than 100 messages, idiot.`, ephemeral: true });
                 const size = await interaction.channel.bulkDelete(amount, true).catch(err => {
